use int32_t constant_pool value

+++ sperl_constant_pool_field.h
@@ -6,7 +6,7 @@
 // Array
 struct SPerl_constant_pool_field {
   int32_t abs_name_constant_pool_address;
-  uint16_t package_byte_offset;
+  int32_t package_byte_offset;
 };
 
 #endif
